backend/requirements.txt
# Core FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.20

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Database clients
pymongo==4.10.1
pinecone==5.3.1

# AI and ML libraries - pinned for stability
langchain==0.3.7
langchain-google-genai==2.0.5
google-generativeai==0.8.3
numpy==1.26.4

# File processing
python-docx==1.1.2
PyPDF2==3.0.1
pandas==2.2.3
openpyxl==3.1.5

# Image processing and OCR
pytesseract==0.3.13
Pillow==10.4.0

# Cloud storage
boto3==1.35.0
google-cloud-storage==2.10.0
google-cloud-language==2.9.1

# Visualization and charts
matplotlib==3.8.4
seaborn==0.13.2

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
email-validator==2.1.1
httpx==0.27.2